Mod reinstall select version
